@@ -29,43 +29,43 @@
     "deploy": "gh-pages -d build"
   },
   "dependencies": {
-    "@reduxjs/toolkit": "^1.9.7",
-    "@types/react": "^16.14.56",
-    "@types/react-dom": "^16.9.24",
+    "@reduxjs/toolkit": "^2.1.0",
+    "@types/react": "^18.2.52",
+    "@types/react-dom": "^18.2.18",
     "@types/react-redux": "^7.1.33",
     "axios": "^1.6.7",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.9",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^9.1.0",
     "react-router-dom": "^6.22.0",
     "react-scripts": "^5.0.1",
     "sass": "^1.70.0",
-    "typescript": "~4.1.5"
+    "typescript": "~5.3.3"
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^12.1.5",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.5.2",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
+    "@testing-library/jest-dom": "^6.4.1",
+    "@testing-library/react": "^14.2.1",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.12",
+    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/parser": "^6.20.0",
     "eslint": "^8.56.0",
-    "eslint-config-prettier": "^8.10.0",
+    "eslint-config-prettier": "^9.1.0",
     "eslint-config-standard": "^17.1.0",
     "eslint-plugin-filename-rules": "^1.3.1",
     "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jest-dom": "^4.0.3",
-    "eslint-plugin-n": "^15.7.0",
+    "eslint-plugin-jest-dom": "^5.1.0",
+    "eslint-plugin-n": "^16.6.2",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-unicorn": "^50.0.1",
-    "gh-pages": "^4.0.0",
-    "husky": "^7.0.4",
+    "gh-pages": "^6.1.1",
+    "husky": "^9.0.10",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.8.8",
+    "prettier": "^3.2.5",
     "serve": "^14.2.1"
   },
   "jest": {
